# docs/openapi/ratings.json
{
  "components": {
    "schemas": {
      "SafeErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": ["code", "message", "requestId"],
            "type": "object"
          }
        },
        "required": ["error"],
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessToken": {
        "description": "Current opaque owner session. Receipts never grant access to content.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "WhaleU independent ratings and text roots",
    "version": "1"
  },
  "openapi": "3.0.3",
  "paths": {
    "/v1/ratings/categories": {
      "get": {
        "description": "Strict current-authorized R1 contract. Independent score and text; unknown authority fails closed.",
        "operationId": "ratingCategories",
        "parameters": [
          {
            "in": "query",
            "name": "regionId",
            "required": false,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parentId",
            "required": false,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^[1-9][0-9]?$",
                  "type": "string"
                }
              ]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{43}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "context": {
                      "additionalProperties": false,
                      "properties": {
                        "catalogRevision": {
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        },
                        "parentId": {
                          "format": "uuid",
                          "nullable": true,
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        },
                        "regionId": {
                          "format": "uuid",
                          "nullable": true,
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                          "type": "string"
                        }
                      },
                      "required": ["regionId", "catalogRevision", "parentId"],
                      "type": "object"
                    },
                    "continuation": {
                      "enum": ["more", "scan", "end"],
                      "type": "string"
                    },
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "description": {
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "kind": {
                            "pattern": "^[a-z][a-z0-9_]{0,49}$",
                            "type": "string"
                          },
                          "level": {
                            "anyOf": [
                              {
                                "enum": [1],
                                "type": "number"
                              },
                              {
                                "enum": [2],
                                "type": "number"
                              },
                              {
                                "enum": [3],
                                "type": "number"
                              }
                            ]
                          },
                          "name": {
                            "type": "string"
                          },
                          "parentId": {
                            "format": "uuid",
                            "nullable": true,
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "revision": {
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                            "type": "string"
                          },
                          "systemKey": {
                            "nullable": true,
                            "pattern": "^[a-z][a-z0-9_]{1,48}$",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "parentId",
                          "level",
                          "kind",
                          "systemKey",
                          "name",
                          "description",
                          "revision"
                        ],
                        "type": "object"
                      },
                      "maxItems": 50,
                      "type": "array"
                    },
                    "nextCursor": {
                      "nullable": true,
                      "pattern": "^[A-Za-z0-9_-]{43}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "context",
                    "items",
                    "nextCursor",
                    "continuation"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Strict ratings contract",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "retry-after": {
                "schema": {
                  "enum": ["60"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            },
            "description": "Strict, current-authorized, sanitized error. Unknown review/verification/feature facts fail closed. Receipts never disclose current private fields.",
            "headers": {
              "cache-control": {
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "vary": {
                "schema": {
                  "enum": ["Authorization"],
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "",
        "tags": ["Ratings"]
      }
    },
...
  },
  "servers": [],
...
}